Release v1.0.0
This is intended to help potential users of this library assess the API
stability and general maturity of this library.

v1.0.0 is a commitment that the API will not change in a backwards
incompatible way before v2.0.0.  The API is stable and has been since the
crate was made available.  In fact the implementation hasn't had to change
in 2 years.  It is time for v1, and the API commitment that comes with it.